New Feature: Additional files to serve can be added to a running instance of SimplePost!
The SimplePost class has had support for adding new files to the server since the first release,
but this commit adds it to the front end program as well. The syntax is documented in the Help()
function. Basically it just involves starting an instance of SimplePost, getting its PID, then
launching SimplePost again with the switch '--pid=PID' (where PID is the process identifier of the
other instance of SimplePost you launched). Adding a file to the first instance of SimplePost
running on your system might look something like this:
./simplepost --pid=$(ps -A | grep simplepost | head -n 1 | awk '{print $1}') --count=7 INSTALL

The technical implementation of this feature involves named pipes. The instance you are adding
a file to (hereby the 'daemon') is sent a signal by another instance (hereby the 'client') which
instructs the daemon to create a named pipe in a predictable location using its PID to
differentiate its pipe from potential other instances. The client waits a predetermined period of
time (defined as SP_PIPE_DELAY at the top of src/simplemain.cpp) before connecting to this pipe
and initiating communication. When the client sends the appropriate command the daemon safely shuts
down the pipe.

There are a descent number of sanity checks built into this initial implementation, but it is a
descently complex feature relative to the size of this program, so there may be bugs.