Added support for specifying more than one file to serve simultaneously.
The SimplePost class has had the capacity to serve more than one file since the first release, but
I didn't implement that functionality on the front end because I didn't need it at the time. This
feature exposes that functionality to the user. It seemed like a hack to require the user to use
the feature introduced in the last commit to serve more than one file at once; therefore I
implemented this feature. Incidentally this feature works in conjunction with the last one - no
additional changes required.

There is one major limitation imposed by this implementation: each file specified on the command
line must be served the same number of times. The back end supports an individual count for each
file, but that required more work to implement on the front end. However, it shouldn't be too much
work to implement at some point because most of the support is already in place, just not exposed.
This limitation can be worked around by adding additional files to serve using the '--pid' switch.