Updated the code for modern compilers
Unfortunately I have not touched this code for a little over 6 months, and in
that relatively short period bitrot set in. I was running Debian Wheezy (which
was Testing) when I wrote this program, and it compiled perfectly. However I
tried compiling it today on Debian Jessie (currently Testing), and both g++
4.8 and clang++ 3.4 refused. Apparently g++ 4.7 and clang++ 3.0 were less
strict with certain includes (or was it eglibc 2.13?). In any case, it was
completely my fault; I should have followed the standard more closely. This
commit fixes the build failure described above.