Significantly refactored makefiles
The old makefiles for this project were based on a template I wrote years ago.
Frankly, they were a complete mess. Since I just finished writing a much
cleaner makefile for an embedded project, I decided to refactor the makefiles
for this project as well. The basic usage is still the same, but the makefiles
are much cleaner internally and should be easier to maintain in the future. As
an added benefit, the 'clean' target now does what you would expect.