Implemented a new interprocess communication handler
The pipe communication mechanism introduced in this commit is a C rewrite
based on my former C++ code, but, more importantly, it also represents a major
improvement in robustness, code quality, and functionality. The internal
protocol implemented in this rewrite IS NOT COMPATIBLE with the old
(SimplePost 1.2.x) implementation. If either one tries to interact with the
other, they will both choke and die a painful death. I have no intention of
fixing this. This is a breaking change, but it hopefully lays the groundwork
for a more stable pipe command infrastructure for more compatability between
future releases.

The implementation in this commit is not quite complete. It compiles, but it
has not been tested because it is missing an implementation of some of the
pipe commands it is supposed to respond to. The internal SimplePost web server
needs to be reimplemented in C before this pipe infrastructure can be
completed.