Initial refactor of the SimplePost web server into C
The major web server changes introduced in a commit are essentially a copy of
commit ce3285ae0f3150b60dcbb8ea1d45b716dfccd116. It was originally introduced
on my libsimplepost branch as part of an effort to both port SimplePost to C
and refactor the web server into its own library. The version in this commit
contains only the essential changes to use it within the infrastructure I've
already restructured.

If the SimplePost C port in this branch succeeds, the libsimplepost branch
will probably go away (without ever being merged back into master). Therefore
the substance of the original commit message - as it relates to the actual
contents of this commit - is documented below.

The former SimplePost C++ class has been rewritten in C with a more-or-less
equivalent (but slightly modified) interface. The behavior of some parts of
the public interface was intetionally altered to fix design flaws I perceived
in the original interface. The goal is for SimplePost 0.3.0 to introduce a
stable API. In the process of this conversion most of the internal interface
was substantially refactored (or in some cases completely rewritten). As a
result this new code most likely has some pretty major bugs.