Downgrade unused warnings/errors
In some cases unused variables, parameters, functions, and results are
desirable (or necessary). This commit modifies the master makefile to instruct
the compiler to only print warnings for unused variables when compiling in
release mode, and to downgrade unused parameters, functions, and results from
errors to warnings when compiling in debug mode.