Added header strings to impact component output
I added header strings (in the form of SimplePost::ComponentName) to the
standard output and standard error messages printed by each major component of
this program. They should make it easier to identify which component a message
came from. This can be useful for both debugging and end users (assuming that
each component is logically broken out, which I would certainly like to
believe). I also added a few new debug messages and improved some of the
existing messages.