Reintroduced the magic file Content-Type header
The last revision of simplepost before the libmicrohttpd transition contained
support for sending the Content-Type header (which HTTP/1.1 strongly recommends
be sent before serving a file). I implemented support for Content-Type by using
libmagic to determine the MIME type of the file and sending the Content-Type
header if the MIME type could be determined. Although the aforementioned
approach hasn't yet been tested in a stable version of this program, it is
theoretically immensely better than the method I used in SimplePost 0.2.x
(which involved trying to determine the MIME type based on the file's extension
alone).

This commit reintroduces Content-Type support using almost exactly the same
method as in commit e0df138d24d761800ba8c79132cddd01457c3ae3, which is
described above.