Disable optimizations by default when configured with --enable-debug
By default the configure script enables "-O2" for GCC. While this is a very
good, safe choice to generate a faster, more optimized executable in release
mode, it makes it kind of annoying stepping through the code with a debugger.
Therefore this commit explicitly disables optimizations when the
"--enable-debug" switch is passed to the configure script.

If you really are trying to debug something with an optimized binary (because
there is a bug that is only reproducible with optimizations enabled, or you
really need the speed, or something), that can still be done by explicitly
setting the "CFLAGS" environment variable before calling configure. For
example, './configure --enable-debug CFLAGS="-O2 -g"'.