Allow SimplePost instances to be listed by address and port
When I first implemented the feature to allow the files being served by other
SimplePost instances to be listed, the PID of the other instance was required
to be specified from the command line. However when adding files to another
instance, that instance could be selected by PID, ADDRESS, PORT, or any
combination of the three. This commit brings the file list option in line
with that behavior, making it not only more consistent, but more convenient
to use.