Process all files, even if the "-c" switch is specified
This was a crazy bug. I noticed while testing something completely unrelated
that if I ran "simplepost -p 5555 README COPYING -c 1 NEWS TODO", the README
and COPYING files would be skipped - as if the argument parser didn't even
see them - but the NEWS and TODO files would be properly served, with their
correct, respective counts. However if I passed the "-c" argument for either
the first file or no files, all files would be processed correctly.

This commit fixes the aforementioned bug, although I am still not completely
sure whether it stemmed from my inexcusable abuse of getopt(), and I simply
worked around it, or if I found a legitimate bug in the glibc 2.22
implementation of getopt(). I am leaning towards the latter, but I am not
sure enough to potentially waste a ton of my time trying to prove it and file
a bug report upstream.