Raise the number of files which may be served simultaneously to SIZE_MAX
This commit removes a completely arbitrary restriction on the number of files
that can be served by the web server at once. Although I remember limiting it
to serving 50 files at a time, that restriction has long since ceased to be
relevant.

When I rewrote SimplePost in C and replaced its internal HTTP server with
libmicrohttpd (in version 0.3.0), it suddenly started scaling much better -
probably a combination of the new architecture I introduced in the rewrite and
the much more robust HTTP server library that I rebased it on.

Now the only limit on the number of files which may be served simultaneously
is the number that simplepost can count: SIZE_MAX. Note that since size_t can
store the largest unsigned integer that is efficient for the platform, the
file limit will be larger on most 64-bit platforms (such as amd64 and arm64)
than on most 32-bit platforms (such as i686 or armv7l). I believe this new
limit is reasonably practical (although I reserve the right to contradict
myself later).