Allow the number of times that a file will be served to be changed
Prior to this commit, once a file was added to the web server, the number of
times that it would be served could only be decremented, never incremented.
The only way to allow a file that was not being hosted indefinitely to be
served more than its remaining count was to download it that many times, then
add it again. Obviously that was not ideal, so this commit introduces the
capability for the number of times that a file will be served to be changed
(either higher or lower) by trying to add a file with the same URI (which
practically speaking means a file with the same name, for the moment anyway).

This new feature also introduces a bug fix: no file will be added more than
once. Technically that never should have happened before - although SimplePost
should have rejected the duplicate and exited with an error before, whereas
now it will change the properties of the file with the same URI - but the
validation code that was supposed to detect duplicates did not always work.