Add a new configure flag to enable argument parser debug messages
This commit adds the "--enable-debug-arg" flag to the configure script to
enable additional messages that are very useful for debugging the command line
argument parser. Prior to this commit these messages were compiled into the
executable every time debugging support was requested. Although I did not want
to lose the ability to debug the command line argument parser, these extra
messages made it somewhat difficult to see other messages, and they are
frankly annoying when I am not debugging the argument parser. Therefore they
must now be explicitly enabled through the configure script with both the
"--enable-debug" and "--enable-debug-arg" flags. (Since this is a debug
feature, the latter will not be honored without the former.)