Bug fix: Always send the simplecmd URL field last
This is a fix for a very subtle bug that I noticed while testing something
else: the COUNT field was ignored when running "simplepost -lf" to list all
files in another instance. It always showed the last file as being served
"indefinitely", but the COUNT for all other files was correct.

A good explanation of the fix is commented in this patch to hopefully make
sure that no one accidentally reintroduces it.