Require ax_prog_doxygen.m4 v14 or newer
This patch allows the configure script for SimplePost to be built with
versions of autoconf-archive newer than November 23, 2015, and removes
support for building with older versions of the m4 macro file. The new
maintainer of ax_prog_doxygen.m4, Olaf Mandel, committed a patch that
replaced the old DX_COND_* conditionals set by autoconf with a new
DX_RULES substitution, which requires backwards-incompatible changes to our
automake files (but *is* much nicer and more future-proof than the old way),
on that day.

I realize that this has been a long time comming, but now that the latest
Debian, Ubuntu, and Fedora releases all have the requisite version of
autoconf-archive packaged, this seems like a good time to switch to the new
ax_prog_doxygen.m4 version. While this patch unfortunately drops support for
older distro releases like Ubuntu 14.04 and RHEL 6, it finally adds support
for much more up-to-date ones like Debian Sid and Arch Linux.