Download a new m4 macro file by default if the version doesn't match
Prior to this commit, the bootstrap script would automatically download a
required m4 macro file from the Internet if it could not be found on the
system (and it wasn't explicitly disallowed with the "-n" switch). This
allowed distros that don't ship autoconf-archive but *do* have new enough
versions of autoconf and automake to bootstrap the configure script. Then I
introduced version checking for certain macro files, and if the file was
found but the version was too old, bootstrap would fail. This patch enables
the same functionality if the version does not match: the bootstrap script
will try to automatically download a new version of the m4 macro file if it
can.

Since the version of autoconf-archive shipped with Debian Jessie and Ubuntu
16.04 is too old for us to use, this patch allows us to bootstrap on those
distros once again.