Fix stringop-overflow warning emitted by GCC 8.2
This commit fixes the "simplepost.c:448:2: warning: ‘strncpy’ specified
bound depends on the length of the source argument
[-Wstringop-overflow=]" warning emitted by GCC 8.2, but it does so in a
rather ham-fisted way. As best I can tell, there was no problem in the
first place. GCC detected that the length of the source string passed
into strncpy() in __get_default_address() was computed using strlen(),
which is the truth. However, we have already made sure that the
destination string is large enough to receive the source by the time we
get to the strcpy(), so there is no possibility of a buffer overflow,
the thing that the compiler's warning message was clearly meant to warn
against. Therefore I fixed this warning in the simplest way possible, and
just replaced the strncpy() with a strcpy().