Fix the format warning about unsigned short emitted by Clang 6.0.1
This commit fixes the "format specifies type 'unsigned short' but the
argument has type 'int' [-Wformat]" warning emitted by Clang 6.0.1. As
best I can tell, this warning is a result of the fact that Clang's
USHRT_MAX definition is actually larger than the largest unsigned
integer that can fit in an unsigned short. It seems like a bug in the
headers that ship with the compiler (as lothe as I usually am to assign
blame to the compiler), but I'm not completely sure of that. (GCC 8.2
with strict format security enabled is perfectly fine with it as-is, for
whatever that's worth.) In any case, it is pretty easy to work around
with a simple cast, and seems to have no functional impact.